+++ RELEASES.md
@@ -3,6 +3,7 @@
 Released on TBD (UTC)
 
 ### Enhancements
+- Added `ActorExecutor` with `Actor` API for creating and running threaded tasks in live environments
 - Implemented string interning for `TimeEvent`
 
 ### Breaking Changes